+++ devices/korg_minilogue_xd/__init__.py
@@ -1,0 +1,1 @@
+"""Korg minilogue xd device helpers."""